test(harness): move long-tail live chat tests to chat_live_longtail suite
One suite of env-gated cells per audit 8c:
- moved whole: test_mistral_api.py, test_openrouter.py, test_snowflake.py,
  test_a2a.py
- test_cohere.py: the 12 live v1/v2 chat/citations/tool tests; the 9 live
  embed v4 tests stay (non-chat, Sameer)
- test_groq.py: TestGroq; structured-output and chunk_parser goldens stay
- test_together_ai.py: TestTogetherAI; supported-params golden stays
- test_xai.py: TestXAIChat, TestXAIReasoningEffort,
  test_xai_message_name_filtering, test_xai_streaming_with_include_usage;
  5 param-mapping goldens stay
- test_deepseek_completion.py: TestDeepSeekChatCompletion (@skip, API
  hanging), test_completion_cost_deepseek; fill_reasoning_content goldens stay
- test_lambda_ai.py / test_v0.py: the env-gated live completion call each;
  config/registry goldens stay
- test_langgraph.py: 2 live local-server tests; config transforms stay
