+++ tests/harness_suites/chat_live_longtail/test_mistral_api.py
@@ -28,7 +28,6 @@ from httpx import Headers
 from base_llm_unit_tests import BaseLLMChatTest
 
 
-@pytest.mark.flaky(retries=3, delay=2)
 class TestMistralCompletion(BaseLLMChatTest):
     def get_base_completion_call_args(self) -> dict:
         litellm.set_verbose = True
