+++ tests/harness_suites/chat_live_longtail/test_together_ai_live.py
@@ -1,0 +1,17 @@
+"""
+Test TogetherAI LLM
+"""
+
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath("../.."))
+
+import litellm
+from base_llm_unit_tests import BaseLLMChatTest
+
+
+class TestTogetherAI(BaseLLMChatTest):
+    def get_base_completion_call_args(self) -> dict:
+        litellm.set_verbose = True
+        return {"model": "together_ai/Qwen/Qwen2.5-7B-Instruct-Turbo"}
